muiをNext.jsのAppディレクトリに対応させる
