+++ app/page.tsx
@@ -1,13 +1,14 @@
+'use client'
+
 import * as React from 'react'
 import Container from '@mui/material/Container'
 import Typography from '@mui/material/Typography'
 import Box from '@mui/material/Box'
-import Button from '@mui/material/Button'
-import Link from '~/components/Link'
-import ProTip from '~/components/ProTip'
-import Copyright from '~/components/Copyright'
+import Link from '~/src/components/Link'
+import ProTip from '~/src/components/ProTip'
+import Copyright from '~/src/components/Copyright'
 
-export default function About() {
+export default function Home() {
   return (
     <Container maxWidth="lg">
       <Box
@@ -22,11 +23,9 @@ export default function About() {
         <Typography variant="h4" component="h1" gutterBottom>
           MUI v5 + Next.js with TypeScript example
         </Typography>
-        <Box maxWidth="sm">
-          <Button variant="contained" component={Link} noLinkStyle href="/">
-            Go to the home page
-          </Button>
-        </Box>
+        <Link href="/about" color="secondary">
+          Go to the about page
+        </Link>
         <ProTip />
         <Copyright />
       </Box>
