--- src/app/head.tsx
@@ -1,9 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title></title>
-      <meta content="width=device-width, initial-scale=1" name="viewport" />
-      <link rel="icon" href="/favicon.ico" />
-    </>
-  )
-}
