--- src/app/layout.tsx
@@ -1,8 +1,0 @@
-export default function RootLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <html>
-      <head />
-      <body>{children}</body>
-    </html>
-  )
-}
