+++ src/components/Copyright.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import * as React from 'react'
 import Typography from '@mui/material/Typography'
 import MuiLink from '@mui/material/Link'
