+++ src/components/Link.tsx
@@ -1,6 +1,8 @@
+'use client'
+
 import * as React from 'react'
 import clsx from 'clsx'
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/navigation'
 import NextLink, { LinkProps as NextLinkProps } from 'next/link'
 import MuiLink, { LinkProps as MuiLinkProps } from '@mui/material/Link'
 import { styled } from '@mui/material/styles'
