+++ tsconfig.json
@@ -17,7 +17,8 @@
     "incremental": true,
     "baseUrl": "./",
     "paths": {
-      "~/*": ["src/*"]
+      "~/app/*": ["app/*"],
+      "~/src/*": ["src/*"]
     },
     "plugins": [
       {
